+++ Data/Apart/Apart.hs
@@ -5,6 +5,7 @@ import Data.Bifoldable (Bifoldable (..))
 import Data.Bifunctor (Bifunctor (..))
 import Data.Bitraversable (Bitraversable (..))
 import Data.Functor.Apply (Apply (..))
+import Data.Functor.Compose (Compose)
 import Data.Kind (Type)
 
 import Data.Apart.Shape (Shape (..))
@@ -38,6 +39,11 @@ instance Traversable t => Bitraversable (Apart t) where
 
 type family Segment (structure :: Type -> Type) (value :: Type) :: Type where
 	Segment (Cofree t) value = t (Cofree t value)
+
+type (:.:) = Compose
+
+type family Injected (structure :: Type -> Type) (extension :: Type -> Type) (value :: Type) :: Type where
+	Injected (Cofree t) extension value = Cofree (t :.: extension) value
 
 type family Scattered (structure :: Type -> Type) (value :: Type) (raw :: Type) :: Type where
 	Scattered (Cofree t) value raw = Apart t raw value