+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 # 0.1.2
 * Add Red-black tree definition and functions
+* Get rid Stack in favor new Nix-style Cabal